requirements.txt
llama-index
nltk
sentence-transformers
faiss-cpu
transformers
